coin-selfservice-war/src/main/java/nl/surfnet/coin/selfservice/control/user/ServiceListController.java
/*
 * Copyright 2012 SURFnet bv, The Netherlands
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package nl.surfnet.coin.selfservice.control.user;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

import javax.annotation.Resource;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.ModelAttribute;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.servlet.ModelAndView;

import nl.surfnet.coin.selfservice.control.BaseController;
import nl.surfnet.coin.selfservice.domain.IdentityProvider;
import nl.surfnet.coin.selfservice.domain.ServiceProvider;
import nl.surfnet.coin.selfservice.service.ServiceProviderService;

/**
 * Controller for listing all service providers for the end user
 */
@Controller
@RequestMapping(value = "/user/*")
public class ServiceListController extends BaseController {

  @Resource(name = "providerService")
  private ServiceProviderService providerService;

  @RequestMapping(value = "linked-services")
  public ModelAndView listLinkedSps(@ModelAttribute(value = "selectedidp") IdentityProvider selectedidp) {
    Map<String, Object> m = new HashMap<String, Object>();
    final List<ServiceProvider> providers = providerService.getLinkedServiceProviders(selectedidp.getId());
    m.put("sps", providers);
    m.put("activeSection", "linked-services");
    return new ModelAndView("user/service-overview", m);
  }
}
